pyproject.toml
[project]
name = "ideas_app"
version = "0.1.0"
description = "Social network app to create and share ideas"
authors = [
    {name = "Javier Hernandez"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "strawberry-graphql-django (>=0.55.1,<0.56.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "strawberry-django-auth (>=0.378.3,<0.379.0)",
    "django-choices-field (>=2.3.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"